[frontend] refacto related analyses on observables - move dummy loading list in its own module
