+++ opencti-platform/opencti-front/src/private/components/observations/stix_cyber_observables/StixCyberObservableMalwareAnalyses.tsx
@@ -2,7 +2,6 @@ import React, { FunctionComponent } from 'react';
 import { graphql, PreloadedQuery, useFragment, usePreloadedQuery } from 'react-relay';
 import List from '@mui/material/List';
 import ListItemIcon from '@mui/material/ListItemIcon';
-import Skeleton from '@mui/material/Skeleton';
 import ListItemText from '@mui/material/ListItemText';
 import ListItem from '@mui/material/ListItem';
 import makeStyles from '@mui/styles/makeStyles';
@@ -26,6 +25,7 @@ import {
 import {
   StixCyberObservableMalwareAnalyses_data$key,
 } from './__generated__/StixCyberObservableMalwareAnalyses_data.graphql';
+import StixCyberObservableMalwareAnalysesDummyList from './StixCyberObservableMalwareAnalysesDummyList';
 
 const useStyles = makeStyles<Theme>((theme) => ({
   goIcon: {
@@ -91,14 +91,6 @@ const useStyles = makeStyles<Theme>((theme) => ({
     marginTop: -15,
   },
 }));
-
-interface StixCyberObservableMalwareAnalysesComponentProps {
-  observableId: string
-}
-
-interface StixCyberObservableMalwareAnalysesListProps {
-  queryRef: PreloadedQuery<StixCyberObservableMalwareAnalysesQuery>
-}
 
 const query = graphql`
   query StixCyberObservableMalwareAnalysesQuery(
@@ -195,6 +187,10 @@ const StixCyberObservableMalwareAnalysesListLine: FunctionComponent<StixCyberObs
   );
 };
 
+interface StixCyberObservableMalwareAnalysesListProps {
+  queryRef: PreloadedQuery<StixCyberObservableMalwareAnalysesQuery>
+}
+
 const StixCyberObservableMalwareAnalysesList: FunctionComponent<StixCyberObservableMalwareAnalysesListProps> = ({ queryRef }) => {
   const { t } = useFormatter();
   const classes = useStyles();
@@ -222,49 +218,9 @@ const StixCyberObservableMalwareAnalysesList: FunctionComponent<StixCyberObserva
   );
 };
 
-const DummyStixCyberObservableMalwareAnalysesList: FunctionComponent = () => {
-  const classes = useStyles();
-
-  return (
-    <>{Array(3)
-      .fill(0)
-      .map((index) => (
-        <ListItem
-          key={index}
-          dense={true}
-          divider={true}
-        >
-          <ListItemIcon classes={{ root: classes.itemIcon }}>
-            <Skeleton
-              animation="wave"
-              variant="circular"
-              width={30}
-              height={30}
-            />
-          </ListItemIcon>
-          <ListItemText
-            primary={
-              <Skeleton
-                animation="wave"
-                variant="rectangular"
-                width="90%"
-                height={15}
-                style={{ marginBottom: 10 }}
-              />
-            }
-            secondary={
-              <Skeleton
-                animation="wave"
-                variant="rectangular"
-                width="90%"
-                height={15}
-              />
-            }
-          />
-        </ListItem>
-      ))}</>
-  );
-};
+interface StixCyberObservableMalwareAnalysesComponentProps {
+  observableId: string
+}
 
 const StixCyberObservableMalwareAnalyses: FunctionComponent<StixCyberObservableMalwareAnalysesComponentProps> = ({ observableId }) => {
   const { t } = useFormatter();
@@ -289,7 +245,7 @@ const StixCyberObservableMalwareAnalyses: FunctionComponent<StixCyberObservableM
       </Typography>
       <List className={classes.list}>
         {queryRef && (
-          <React.Suspense fallback={<DummyStixCyberObservableMalwareAnalysesList/>}>
+          <React.Suspense fallback={<StixCyberObservableMalwareAnalysesDummyList/>}>
             <StixCyberObservableMalwareAnalysesList queryRef={queryRef}/>
           </React.Suspense>
         )}
